+++ ndk/lib/src/main/java/io/sentry/ndk/INativeScope.java
@@ -15,4 +15,6 @@ public interface INativeScope {
 
   void addBreadcrumb(
       String level, String message, String category, String type, String timestamp, String data);
+
+  void setTrace(String traceId, String parentSpanId);
 }