+++ src/sentry_sync.c
@@ -9,7 +9,7 @@
 #ifdef SENTRY_PLATFORM_WINDOWS
 typedef HRESULT(WINAPI *pSetThreadDescription)(
     HANDLE hThread, PCWSTR lpThreadDescription);
-const DWORD MS_VC_EXCEPTION = 0x406D1388;
+static const DWORD MS_VC_EXCEPTION = 0x406D1388;
 
 #    pragma pack(push, 8)
 typedef struct {
@@ -20,14 +20,14 @@ typedef struct {
 } THREADNAME_INFO;
 #    pragma pack(pop)
 
-sentry_threadid_t
-sentry__thread_get_current_threadid(void)
+static sentry_threadid_t
+thread_get_current_threadid(void)
 {
     return GetCurrentThread();
 }
 
-int
-sentry__thread_setname(sentry_threadid_t thread_id, const char *thread_name)
+static int
+thread_setname(sentry_threadid_t thread_id, const char *thread_name)
 {
     if (!thread_id || !thread_name) {
         return 0;
@@ -52,7 +52,10 @@ sentry__thread_setname(sentry_threadid_t thread_id, const char *thread_name)
     threadnameInfo.dwThreadID
         = GetThreadId(thread_id); // only available on Windows Vista+
     threadnameInfo.dwFlags = 0;
-
+#        ifdef __clang__
+#            pragma clang diagnostic push
+#            pragma clang diagnostic ignored "-Wlanguage-extension-token"
+#        endif
 #        pragma warning(push)
 #        pragma warning(disable : 6320 6322)
     __try {
@@ -62,19 +65,22 @@ sentry__thread_setname(sentry_threadid_t thread_id, const char *thread_name)
     } __except (EXCEPTION_EXECUTE_HANDLER) {
     }
 #        pragma warning(pop)
+#        ifdef __clang__
+#            pragma clang diagnostic pop
+#        endif
 #    endif
 
     return 0;
 }
 #else
-sentry_threadid_t
-sentry__thread_get_current_threadid(void)
+static sentry_threadid_t
+thread_get_current_threadid(void)
 {
     return pthread_self();
 }
 
-int
-sentry__thread_setname(sentry_threadid_t thread_id, const char *thread_name)
+static int
+thread_setname(sentry_threadid_t thread_id, const char *thread_name)
 {
     if (!thread_id || !thread_name) {
         return 0;
@@ -234,8 +240,7 @@ worker_thread(void *data)
     // subject to initialization race condition: current thread might be running
     // before `bgw->thread_id` is initialized in the thread that started the
     // background worker.
-    if (sentry__thread_setname(
-            sentry__thread_get_current_threadid(), bgw->thread_name)) {
+    if (thread_setname(thread_get_current_threadid(), bgw->thread_name)) {
         SENTRY_WARN("failed to set background worker thread name");
     }