+++ src/sentry_utils.c
@@ -63,7 +63,8 @@ sentry__url_parse(sentry_url_t *url_out, const char *url, bool requires_path)
     if (!tmp) {
         goto error;
     }
-    url_out->scheme = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+    url_out->scheme
+        = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
 
     if (!url_out->scheme || !is_scheme_valid(url_out->scheme)) {
         goto error;
@@ -96,14 +97,15 @@ sentry__url_parse(sentry_url_t *url_out, const char *url, bool requires_path)
     tmp = ptr;
     if (has_username) {
         SKIP_WHILE_NOT2(tmp, '@', ':');
-        url_out->username = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+        url_out->username
+            = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
         ptr = tmp;
         if (*ptr == ':') {
             ptr++;
             tmp = ptr;
             SKIP_WHILE_NOT(tmp, '@');
             url_out->password
-                = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+                = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
             ptr = tmp;
         }
         if (*ptr != '@') {
@@ -126,7 +128,7 @@ sentry__url_parse(sentry_url_t *url_out, const char *url, bool requires_path)
         tmp++;
     }
 
-    url_out->host = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+    url_out->host = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
 
     /* port */
     ptr = tmp;
@@ -134,7 +136,7 @@ sentry__url_parse(sentry_url_t *url_out, const char *url, bool requires_path)
         ptr++;
         tmp = ptr;
         SKIP_WHILE_NOT(tmp, '/');
-        aux_buf = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+        aux_buf = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
         char *end;
         url_out->port = (int)strtol(aux_buf, &end, 10);
         if (end != aux_buf + strlen(aux_buf)) {
@@ -169,7 +171,7 @@ sentry__url_parse(sentry_url_t *url_out, const char *url, bool requires_path)
     /* path */
     tmp = ptr;
     SKIP_WHILE_NOT2(tmp, '#', '?');
-    url_out->path = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+    url_out->path = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
     ptr = tmp;
 
     /* query */
@@ -177,7 +179,8 @@ sentry__url_parse(sentry_url_t *url_out, const char *url, bool requires_path)
         ptr++;
         tmp = ptr;
         SKIP_WHILE_NOT(tmp, '#');
-        url_out->query = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+        url_out->query
+            = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
         ptr = tmp;
     }
 
@@ -186,7 +189,8 @@ sentry__url_parse(sentry_url_t *url_out, const char *url, bool requires_path)
         ptr++;
         tmp = ptr;
         SKIP_WHILE_NOT(tmp, 0);
-        url_out->fragment = sentry__string_clone_n_unchecked(ptr, tmp - ptr);
+        url_out->fragment
+            = sentry__string_clone_n_unchecked(ptr, (size_t)(tmp - ptr));
     }
 
     result = 0;
@@ -488,7 +492,7 @@ sentry__iso8601_to_usec(const char *iso)
         return 0;
     }
 
-    return (uint64_t)time * 1000000 + usec;
+    return (uint64_t)time * 1000000 + (uint64_t)usec;
 }
 
 #ifdef SENTRY_PLATFORM_WINDOWS
@@ -501,7 +505,12 @@ sentry__iso8601_to_usec(const char *iso)
 // forwards to `stdtod`, but it does not define `vsnprintf_l` sadly.  This means
 // if Android ever adds locale support in NDK we will have to revisit this code
 // to ensure the C locale is also used there.
-#if !defined(SENTRY_PLATFORM_ANDROID) && !defined(SENTRY_PLATFORM_IOS)
+#if !defined(SENTRY_PLATFORM_ANDROID) && !defined(SENTRY_PLATFORM_IOS)         \
+    && !defined(SENTRY_PLATFORM_AIX)
+#    define HAS_C_LOCALE
+#endif
+
+#ifdef HAS_C_LOCALE
 static sentry__locale_t
 c_locale(void)
 {
@@ -523,8 +532,7 @@ sentry__strtod_c(const char *ptr, char **endptr)
 {
 #ifdef SENTRY_PLATFORM_WINDOWS
     return _strtod_l(ptr, endptr, c_locale());
-#elif defined(SENTRY_PLATFORM_ANDROID) || defined(SENTRY_PLATFORM_IOS)         \
-    || defined(SENTRY_PLATFORM_AIX)
+#elif !defined(HAS_C_LOCALE)
     return strtod(ptr, endptr);
 #else
     return strtod_l(ptr, endptr, c_locale());
@@ -540,8 +548,7 @@ sentry__snprintf_c(char *buf, size_t buf_size, const char *fmt, ...)
     int rv;
 #ifdef SENTRY_PLATFORM_WINDOWS
     rv = _vsnprintf_l(buf, buf_size, fmt, c_locale(), args);
-#elif defined(SENTRY_PLATFORM_ANDROID) || defined(SENTRY_PLATFORM_IOS)         \
-    || defined(SENTRY_PLATFORM_AIX)
+#elif !defined(HAS_C_LOCALE)
     rv = vsnprintf(buf, buf_size, fmt, args);
 #elif defined(SENTRY_PLATFORM_DARWIN)
     rv = vsnprintf_l(buf, buf_size, c_locale(), fmt, args);