+++ tests/cmake.py
@@ -15,6 +15,10 @@ class CMake:
         self.factory = factory
 
     def compile(self, targets, options=None):
+        # We build in tmp for all of our tests. Disable the warning MSVC generates to not clutter the build logs.
+        if sys.platform == "win32":
+            os.environ["IgnoreWarnIntDirInTempDetected"] = "True"
+
         if options is None:
             options = dict()
         key = (
@@ -129,17 +133,23 @@ def cmake(cwd, targets, options=None):
             "cmake",
         ]
 
-    configcmd = cmake.copy()
+    config_cmd = cmake.copy()
+
+    if os.environ.get("VS_GENERATOR_TOOLSET") == "ClangCL":
+        config_cmd.append("-G Visual Studio 17 2022")
+        config_cmd.append("-A x64")
+        config_cmd.append("-T ClangCL")
+
     for key, value in options.items():
-        configcmd.append("-D{}={}".format(key, value))
+        config_cmd.append("-D{}={}".format(key, value))
     if sys.platform == "win32" and os.environ.get("TEST_X86"):
-        configcmd.append("-AWin32")
+        config_cmd.append("-AWin32")
     elif sys.platform == "linux" and os.environ.get("TEST_X86"):
-        configcmd.append("-DSENTRY_BUILD_FORCE32=ON")
+        config_cmd.append("-DSENTRY_BUILD_FORCE32=ON")
     if "asan" in os.environ.get("RUN_ANALYZER", ""):
-        configcmd.append("-DWITH_ASAN_OPTION=ON")
+        config_cmd.append("-DWITH_ASAN_OPTION=ON")
     if "tsan" in os.environ.get("RUN_ANALYZER", ""):
-        configcmd.append("-DWITH_TSAN_OPTION=ON")
+        config_cmd.append("-DWITH_TSAN_OPTION=ON")
 
     # we have to set `-Werror` for this cmake invocation only, otherwise
     # completely unrelated things will break
@@ -153,8 +163,24 @@ def cmake(cwd, targets, options=None):
     if "gcc" in os.environ.get("RUN_ANALYZER", ""):
         cflags.append("-fanalyzer")
     if "llvm-cov" in os.environ.get("RUN_ANALYZER", ""):
-        flags = "-fprofile-instr-generate -fcoverage-mapping"
-        configcmd.append("-DCMAKE_C_FLAGS='{}'".format(flags))
+        if False and os.environ.get("VS_GENERATOR_TOOLSET") == "ClangCL":
+            # for clang-cl in CI we have to use `--coverage` rather than `fprofile-instr-generate` and provide an
+            # architecture-specific profiling library for it work:
+            # TODO: This currently doesn't work due to https://gitlab.kitware.com/cmake/cmake/-/issues/24025
+            #       The issue describes a behavior where generated object-name is specified via `/fo` using a target
+            #       directory, rather than a file-name (this is CMake behavior). While the `clang-cl` suggest that this
+            #       is supported the flag produces `.gcda` and `.gcno` files, which have no relation with the object-
+            #       file and which leads to failure to accumulate coverage data.
+            #       This would have to be fixed in clang-cl: https://github.com/llvm/llvm-project/issues/87304
+            #       Let's leave this in here for posterity, it would be great to get coverage analysis on Windows.
+            flags = "--coverage"
+            profile_lib = "clang_rt.profile-x86_64.lib"
+            config_cmd.append(f"-DCMAKE_EXE_LINKER_FLAGS='{profile_lib}'")
+            config_cmd.append(f"-DCMAKE_SHARED_LINKER_FLAGS='{profile_lib}'")
+            config_cmd.append(f"-DCMAKE_MODULE_LINKER_FLAGS='{profile_lib}'")
+        else:
+            flags = "-fprofile-instr-generate -fcoverage-mapping"
+        config_cmd.append("-DCMAKE_C_FLAGS='{}'".format(flags))
 
         # Since we overwrite `CXXFLAGS` below, we must add the experimental library here for the GHA runner that builds
         # sentry-native with LLVM clang for macOS (to run ASAN on macOS) rather than the version coming with XCode.
@@ -169,17 +195,17 @@ def cmake(cwd, targets, options=None):
                 + " -L/usr/local/opt/llvm@15/lib/c++ -fexperimental-library -Wno-unused-command-line-argument"
             )
 
-        configcmd.append("-DCMAKE_CXX_FLAGS='{}'".format(flags))
+        config_cmd.append("-DCMAKE_CXX_FLAGS='{}'".format(flags))
     if "CMAKE_DEFINES" in os.environ:
-        configcmd.extend(os.environ.get("CMAKE_DEFINES").split())
+        config_cmd.extend(os.environ.get("CMAKE_DEFINES").split())
     env = dict(os.environ)
     env["CFLAGS"] = env["CXXFLAGS"] = " ".join(cflags)
 
-    configcmd.append(source_dir)
+    config_cmd.append(source_dir)
 
-    print("\n{} > {}".format(cwd, " ".join(configcmd)), flush=True)
+    print("\n{} > {}".format(cwd, " ".join(config_cmd)), flush=True)
     try:
-        subprocess.run(configcmd, cwd=cwd, env=env, check=True)
+        subprocess.run(config_cmd, cwd=cwd, env=env, check=True)
     except subprocess.CalledProcessError:
         raise pytest.fail.Exception("cmake configure failed") from None